Assignment6 peer-reveiwed I added some if tests to make sure the input did not go over or under the allowed start/end years. Also added buttons to your html code insted of url links.
you need to refresh the browser everytime to see the pictures. This is due to the cache remembering the last (first)
picture you send in to the browser. To fix this you can make a new picture everytime with a diffrent name
and tell it to show that name. To generate totaly random names a smart way can be to use the variables send in
by user and use them to designvish. I did not add this due to the fact that i would have to re-write most of the code
to make the html,web_visulizer and temperature_CO2_plotter talk togheter.

I also added the t_start and t_end to be = to their lowest and highest value by defult. This would come in handy
when dealing with user input.

Your doc-strings looks really good and are really well writen, same with your readme.md!
your code was easy to read after I read the doc-strings

Well done!